% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloOAuth.R
\name{armadillo.get_credentials}
\alias{armadillo.get_credentials}
\title{Get credentials information}
\usage{
armadillo.get_credentials(server)
}
\arguments{
\item{server}{the URL of the Armadillo server}
}
\value{
The credentials
}
\description{
Get credentials, including refresh, id and access token via device flow auth
NOTE: in order to get refresh token, refresh tokens should be turned on for this armadillo instance on the auth server
}
