% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloOAuth.R
\name{armadillo.get_token}
\alias{armadillo.get_token}
\title{Get ID Token}
\usage{
armadillo.get_token(server)
}
\arguments{
\item{server}{the URL of the Armadillo server}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\details{
This function is defunct. Use \code{\link[=armadillo.get_credentials]{armadillo.get_credentials()}} instead.
}
