% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloOAuth.R
\name{armadillo.refresh_connections}
\alias{armadillo.refresh_connections}
\title{Refresh connections of all armadillo connections}
\usage{
armadillo.refresh_connections(
  server,
  credentials,
  conns,
  connections_name,
  env = .GlobalEnv
)
}
\arguments{
\item{server}{the URL of the Armadillo server}

\item{credentials}{a list containing "refresh_token" and "access_token", can be retrieved using armadillo.get_credentials}

\item{conns}{the armadillo connections object}

\item{connections_name}{the name of the connections object in your environment}

\item{env}{your environment (default: .GlobalEnv)}
}
\value{
A list of credentials containing a new refresh_token and access_token "token" that can be used to retrieved data again
}
\description{
Refreshes a timed out connections using credentials containing the old access token and newest refresh token
Old credentials will be overwritten in the connections object to be able to continue analysis
NOTE: make sure refresh is possible for the armadillo instance on the auth server
}
