% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloOAuth.R
\name{datashield.get.token}
\alias{datashield.get.token}
\title{Get ID Token}
\usage{
datashield.get.token(server)
}
\arguments{
\item{server}{the URL of the Armadillo server}
}
\value{
The ID token string
}
\description{
Get an ID token to log in on an Armadillo server or its MinIO filestore.
}
