% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloOAuth.R
\name{.get_info}
\alias{.get_info}
\title{Get oauth server discovery information}
\usage{
.get_info(armadillo_server)
}
\arguments{
\item{armadillo_server}{url of the Armadillo server}
}
\value{
a dataframe with issuerUrl and clientId
}
\description{
Specifically this method returns the 
- issuer URL
- clientId
}
\keyword{internal}
