% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsAssignTable,MolgenisConnection-method}
\alias{dsAssignTable,MolgenisConnection-method}
\title{Assign a table}
\usage{
\S4method{dsAssignTable}{MolgenisConnection}(
  conn,
  symbol,
  table,
  variables = NULL,
  missings = FALSE,
  identifiers = NULL,
  id.name = NULL,
  async = TRUE
)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} object.}

\item{symbol}{Name of the R symbol.}

\item{table}{Fully qualified name of the table.}

\item{variables}{The variables to load.}

\item{missings}{Not supported}

\item{identifiers}{Not supported}

\item{id.name}{Not supported}

\item{async}{When set, do not wait for the result.}
}
\value{
A \code{\link{MolgenisResult-class}} object.
}
\description{
Assign a MOLGENIS table in the DataSHIELD R session.
}
