% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisDriver.R
\name{dsConnect,MolgenisDriver-method}
\alias{dsConnect,MolgenisDriver-method}
\title{Connect to a Molgenis DataSHIELD service}
\usage{
\S4method{dsConnect}{MolgenisDriver}(
  drv,
  name,
  restore = NULL,
  username = NULL,
  password = NULL,
  token = NULL,
  url = NULL,
  opts = list(),
  ...
)
}
\arguments{
\item{drv}{\code{\link{MolgenisDriver-class}} class object.}

\item{name}{Name of the connection, which must be unique among all the
DataSHIELD connections.}

\item{restore}{Workspace name to be restored in the newly created DataSHIELD
R session.}

\item{username}{The username to authenticate with.}

\item{password}{The password to authenticate with.}

\item{token}{Not yet used.}

\item{url}{URL of the server.}

\item{opts}{Curl options as described by httr (call httr::httr_options()
for details). Can be provided by "Molgenis.opts" option.}

\item{...}{Unused, needed for compatibility with generic.}
}
\value{
A \code{\link{MolgenisConnection-class}} object.
}
\description{
Connect to a Molgenis DataSHIELD service, with provided credentials.
}
