% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsDisconnect,MolgenisConnection-method}
\alias{dsDisconnect,MolgenisConnection-method}
\title{Disconnect from a MOLGENIS DataSHIELD Service}
\usage{
\S4method{dsDisconnect}{MolgenisConnection}(conn, save = NULL)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class
object}

\item{save}{Save the DataSHIELD R session with provided ID (must be a
character string).}
}
\description{
Disconnect from a MOLGENIS DataSHIELD Service and release all R resources. If
a workspace ID is provided, the DataSHIELD R session will be saved before
being destroyed.
}
\seealso{
\code{\link{dsDisconnect}}
}
