% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisResult.R
\name{dsFetch,MolgenisResult-method}
\alias{dsFetch,MolgenisResult-method}
\title{Fetch the result}
\usage{
\S4method{dsFetch}{MolgenisResult}(res)
}
\arguments{
\item{res}{\code{\link{MolgenisResult-class}} object.}
}
\value{
TRUE if table exists.
}
\description{
Fetch the DataSHIELD operation result.
}
