% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisResult.R
\name{dsGetInfo,MolgenisResult-method}
\alias{dsGetInfo,MolgenisResult-method}
\title{Get result info}
\usage{
\S4method{dsGetInfo}{MolgenisResult}(dsObj, ...)
}
\arguments{
\item{dsObj}{\code{\link{MolgenisResult-class}} class object}

\item{...}{Unused, needed for compatibility with generic.}
}
\value{
The result information. This should include the R expression
being executed (`expression`) and if the query is complete
(`{ "status" = "COMPLETED" }`).
}
\description{
Get the information about a command (if still available).
}
