% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsHasTable,MolgenisConnection-method}
\alias{dsHasTable,MolgenisConnection-method}
\title{Verify table exist and can be accessible for performing DataSHIELD
operations.}
\usage{
\S4method{dsHasTable}{MolgenisConnection}(conn, table)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class object.}

\item{table}{The identifier of the table}
}
\value{
TRUE if table exists.
}
\description{
Verify table exist and can be accessible for performing DataSHIELD
operations.
}
