% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsIsAsync,MolgenisConnection-method}
\alias{dsIsAsync,MolgenisConnection-method}
\title{MOLGENIS DataShield Service asynchronous support}
\usage{
\S4method{dsIsAsync}{MolgenisConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class object}
}
\value{
The named list of logicals detailing the asynchronicity support.
}
\description{
List of DataSHIELD operations on which MOLGENIS DataSHIELD Service supports
asynchronicity.
}
\details{
When a \code{\link{DSResult-class}} object is returned on aggregation or
assignment operation, the raw result can be accessed asynchronously, allowing
parallelization of DataSHIELD calls over multpile servers. The returned named
list of logicals will specify if asynchronicity is supported for:
aggregation operation ('aggregate'), table assignment operation
('assignTable'), expression assignment operation ('assignExpr').
}
