% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsListMethods,MolgenisConnection-method}
\alias{dsListMethods,MolgenisConnection-method}
\title{List methods}
\usage{
\S4method{dsListMethods}{MolgenisConnection}(conn, type = "aggregate")
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class object}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\value{
A data.frame with columns: name, type ('aggregate' or 'assign'),
class ('function' or 'script'), value, package, version.
}
\description{
List methods defined in the DataSHIELD configuration.
}
