% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsListPackages,MolgenisConnection-method}
\alias{dsListPackages,MolgenisConnection-method}
\title{List packages}
\usage{
\S4method{dsListPackages}{MolgenisConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class object}
}
\value{
A data.frame with columns: name, version.
}
\description{
List packages with their versions defined in the DataSHIELD configuration.
}
