% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsListSymbols,MolgenisConnection-method}
\alias{dsListSymbols,MolgenisConnection-method}
\title{List R symbols}
\usage{
\S4method{dsListSymbols}{MolgenisConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class object}
}
\value{
A character vector.
}
\description{
List symbols living in the DataSHIELD R session.
}
