% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsListTables,MolgenisConnection-method}
\alias{dsListTables,MolgenisConnection-method}
\title{List MOLGENIS DataSHIELD Service tables}
\usage{
\S4method{dsListTables}{MolgenisConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class object}
}
\value{
The fully qualified names of the tables.
}
\description{
List MOLGENIS DataSHIELD Service tables that may be accessible for performing
DataSHIELD operations.
}
\seealso{
\code{\link{dsListTables}}
}
