% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsListWorkspaces,MolgenisConnection-method}
\alias{dsListWorkspaces,MolgenisConnection-method}
\title{List workspaces}
\usage{
\S4method{dsListWorkspaces}{MolgenisConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class object}
}
\value{
A data.frame with columns: name, lastAccessDate, size, user.
}
\description{
List workspaces saved in the data repository.
}
