% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloOAuth.R,
%   R/DSMolgenisArmadillo-defunct.R
\name{armadillo.get_token}
\alias{armadillo.get_token}
\alias{DSMolgenisArmadillo-defunct}
\title{Defunct functions in package \pkg{DSMolgenisArmadillo}.}
\usage{
armadillo.get_token(server)
}
\description{
The functions listed below are defunct and no longer available.
When possible, alternative functions with similar functionality are
mentioned. Help pages for deprecated functions are available at
\code{help("<function>-defunct")}.
}
\section{\code{armadillo.get_token}}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

For \code{armadillo.get_token}, use \code{\link[=armadillo.get_credentials]{armadillo.get_credentials()}} instead.
}

\keyword{internal}
