% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloConnection.R
\name{dsRestoreWorkspace,ArmadilloConnection-method}
\alias{dsRestoreWorkspace,ArmadilloConnection-method}
\title{Restore workspace}
\usage{
\S4method{dsRestoreWorkspace}{ArmadilloConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\link{ArmadilloConnection-class}} class object}

\item{name}{Name of the workspace.}
}
\description{
This method implementation for the ArmadilloConnection class sends an authenticated POST request
to the Armadillo /load-workspace endpoint. It reloads a previously saved R session state
(workspace) using the provided name as the identifier.
}
