% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsRmWorkspace,MolgenisConnection-method}
\alias{dsRmWorkspace,MolgenisConnection-method}
\title{Remove a workspace}
\usage{
\S4method{dsRmWorkspace}{MolgenisConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class object}

\item{name}{Name of the workspace.}
}
\description{
Remove a workspace on the data repository.
}
