% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\name{dsSaveWorkspace,MolgenisConnection-method}
\alias{dsSaveWorkspace,MolgenisConnection-method}
\title{Save workspace}
\usage{
\S4method{dsSaveWorkspace}{MolgenisConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\link{MolgenisConnection-class}} class object}

\item{name}{Name of the workspace.}
}
\description{
Save workspace on the data repository.
}
