% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisConnection.R
\docType{class}
\name{MolgenisConnection-class}
\alias{MolgenisConnection-class}
\title{Class MolgenisConnection.}
\description{
A Molgenis connection implementing the DataSHIELD Interface (DSI)
\code{\link{DSConnection-class}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the connection}

\item{\code{workspaces}}{The shared workspaces to load tables from}

\item{\code{handle}}{The handle used to connect with the server}

\item{\code{user}}{The username used to authenticate}

\item{\code{cookies}}{The cookies set by the server}
}}

\keyword{internal}
