% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisDriver.R
\docType{class}
\name{MolgenisDriver-class}
\alias{MolgenisDriver-class}
\title{Class MolgenisDriver with constructor molgenis}
\description{
An Molgenis DataSHIELD Service Driver implementing the DataSHIELD Interface
(DSI) \code{\link{DSDriver-class}}. This class should always be initialized
with the \code{\link{molgenis}} function.
It returns a singleton that allows you to connect to Molgenis.
}
\keyword{internal}
