% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisResult.R
\docType{class}
\name{MolgenisResult-class}
\alias{MolgenisResult-class}
\title{Class MolgenisResult.}
\description{
A MOLGENIS result implementing the DataSHIELD Interface (DSI)
\code{\link{DSResult-class}}.
}
\section{Slots}{

\describe{
\item{\code{conn}}{The connection used to create this result}

\item{\code{rval}}{The result}
}}

\keyword{internal}
