% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmadilloOAuth.R
\name{armadillo.get_token}
\alias{armadillo.get_token}
\title{Get ID Token}
\usage{
armadillo.get_token(server)
}
\arguments{
\item{server}{the URL of the Armadillo server}
}
\description{
`r lifecycle::badge("defunct")`
}
\details{
This function is defunct. Use [armadillo.get_credentials()] instead.
}
