% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSMolgenisArmadillo-package.R
\docType{package}
\name{DSMolgenisArmadillo-package}
\alias{DSMolgenisArmadillo}
\alias{DSMolgenisArmadillo-package}
\title{DSMolgenisArmadillo: DataSHIELD implementation for the Armadillo}
\description{
DataSHIELD implementation for the Armadillo DataSHIELD service. Part of the MOLGENIS suite.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/molgenis/molgenis-r-datashield}
  \item \url{https://molgenis.github.io/molgenis-r-datashield}
  \item Report bugs at \url{https://github.com/molgenis/molgenis-r-datashield/issues}
}

}
\author{
\strong{Maintainer}: Sido Haakma \email{s.haakma@rug.nl} (\href{https://orcid.org/0000-0003-1278-9144}{ORCID})

Authors:
\itemize{
  \item Tommy de Boer \email{t.de.boer01@umcg.nl} (\href{https://orcid.org/0000-0002-4492-7225}{ORCID})
  \item Fleur Kelpin \email{f.kelpin@umcg.nl} (\href{https://orcid.org/0000-0003-1527-5130}{ORCID})
}

}
\keyword{internal}
